% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{all_are_long}
\alias{all_are_long}
\title{Lange Dataframes?}
\usage{
all_are_long(x)
}
\arguments{
\item{x}{Liste mit Tiqqles, deren Format ueberprueft wird.}
}
\value{
Logischer Wert, ob alle Tiqqles in Liste \code{x} im langen Format
  vorliegen.
}
\description{
Ueberprueft, ob in einer Liste von Dataframes alle im langen Format
vorliegen. Das lange Format besitzt hierbei drei Spalten: \emph{time}
(Zeitstempel), \emph{signal} (Signalname als Faktor) und \emph{value}
(numerischer Signalwert).
}
\seealso{
Other Format-Checks: 
\code{\link{all_are_valid}()},
\code{\link{all_are_wide}()},
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_valid}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
\keyword{internal}
