% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{all_are_valid}
\alias{all_are_valid}
\title{Valide Dataframes?}
\usage{
all_are_valid(x)
}
\arguments{
\item{x}{Liste mit Tiqqles, deren Format ueberprueft wird.}
}
\value{
Logischer Wert, ob alle Tiqqles in Liste \code{x} valide sind.
}
\description{
Ueberprueft, ob in einer Liste von Dataframes alle lange oder breite Tiqqle
sind und Daten enthalten.
}
\seealso{
Other Format-Checks: 
\code{\link{all_are_long}()},
\code{\link{all_are_wide}()},
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_valid}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
\keyword{internal}
