% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{all_are_wide}
\alias{all_are_wide}
\title{Breite Dataframes?}
\usage{
all_are_wide(x)
}
\arguments{
\item{x}{Liste mit Tiqqles, deren Format ueberprueft wird.}
}
\value{
Logischer Wert, ob alle Tiqqles in Liste \code{x} im breiten Format
  vorliegen.
}
\description{
Ueberprueft, ob in einer Liste von Dataframes alle im breiten Format
vorliegen. Das breite Format besitzt hierbei eine Spalte \emph{time}
(Zeitstempel) und fuer jedes Signal eine weitere numerische Spalte mit dem
Namen des Signals.
}
\seealso{
Other Format-Checks: 
\code{\link{all_are_long}()},
\code{\link{all_are_valid}()},
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_valid}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
\keyword{internal}
