% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{all_days}
\alias{all_days}
\title{Tage in Zeitraum?}
\usage{
all_days(int_start, int_end)
}
\arguments{
\item{int_start}{POSIXct-Zeitstempel mit Startzeitpunkt des Zeitraums.}

\item{int_end}{POSIXct-Zeitstempel mit Endzeitpunkt des Zeitraums.}
}
\value{
Date-Vektor mit den Tagen im Zeitraum.
}
\description{
Bestimmt alle Tage, die in einem bestimmten Zeitraum liegen.
}
