% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{all_weekends}
\alias{all_weekends}
\title{Wochenende?}
\usage{
all_weekends(int_start, int_end, is_truncated = TRUE)
}
\arguments{
\item{int_start}{POSIXct-Zeitstempel mit Startzeitpunkt des Suchzeitraums.}

\item{int_end}{POSIXct-Zeitstempel mit Endzeitpunkt des Suchzeitraums.}

\item{is_truncated}{Logischer Wert, ob Wochenendzeitstempel an den Raendern
des Suchzeitraums beschnitten werden sollen, damit keine Zeitstempel
ausserhalb liegen (Default: \emph{TRUE}).}
}
\value{
Dataframe mit den gefundenen Wochenenden.
}
\description{
Bestimmt innerhalb eines Zeitraums die Anfangs- und Endzeitpunkte der
Wochenenden (Samstag 00:00 Uhr bis Montag 00:00 Uhr).
}
