% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{arrange2}
\alias{arrange2}
\title{Sortiert Dataframe (falls noetig)}
\usage{
arrange2(x, var)
}
\arguments{
\item{x}{Dataframe, der ggf. sortiert wird.}

\item{var}{Symbol oder String der Spalte, nach der ggf. sortiert wird.}
}
\value{
Sortierter Dataframe \code{x}.
}
\description{
Ueberprueft, ob Dataframe bereits nach bestimmter Spalte sortiert ist. Falls
nicht, wird eine Sortierung durchgefuehrt.
}
\details{
Erweiterte Variante von \code{\link{arrange}}.
}
\seealso{
\code{\link{arrange}}
}
\keyword{internal}
