% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiqqle.R
\name{as_long}
\alias{as_long}
\title{Breiten Tiqqle in langes Format}
\usage{
as_long(x, remove_na = FALSE)
}
\arguments{
\item{x}{Tiqqle im breiten Format, der konvertiert wird.}

\item{remove_na}{Logischer Wert, ob die \emph{NA}-Eintraege im resultierenden
langen Tiqqle entfernt werden (Default: \emph{FALSE}, d.h. \emph{NA}-Werte
werden nicht entfernt).}
}
\value{
Tiqqle im langen Format.
}
\description{
Ein Dataframe wird vom breiten ins lange Format umgewandelt. Das breite
Format besitzt hierbei eine Spalte \emph{time} (Zeitstempel) und fuer jedes
Signal eine weitere numerische Spalte mit dem Namen des Signals. Das
resultierende lange Format besitzt drei Spalten: \emph{time} (Zeitstempel),
\emph{signal} (Signalname) und \emph{value} (Signalwert).
}
\seealso{
Other Tiqqle-Funktionen: 
\code{\link{as_tiqqle}()},
\code{\link{as_wide}()},
\code{\link{tiqqle}()}
}
\concept{Tiqqle-Funktionen}
