% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiqqle.R
\name{as_tiqqle}
\alias{as_tiqqle}
\title{Dataframe zu Tiqqle}
\usage{
as_tiqqle(x, format = "long")
}
\arguments{
\item{x}{Dataframe, der konvertiert wird.}

\item{format}{String, welches Format der konvertierte Tiqqle hat. Erlaubte
Werte: \emph{"long"} oder \emph{"wide"} (Default: \emph{"long"}, d.h.
langes Format).}
}
\value{
Tiqqle im gewuenschten Format.
}
\description{
Wandelt einen bestehenden Dataframe in einen Tiqqle um.
}
\seealso{
Other Tiqqle-Funktionen: 
\code{\link{as_long}()},
\code{\link{as_wide}()},
\code{\link{tiqqle}()}
}
\concept{Tiqqle-Funktionen}
