% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiqqle.R
\name{as_wide}
\alias{as_wide}
\title{Langen Tiqqle in breites Format}
\usage{
as_wide(x, fill_gap = FALSE)
}
\arguments{
\item{x}{Tiqqle im langen Format, der konvertiert wird.}

\item{fill_gap}{Logischer Wert, ob die \emph{NA}-Luecken im resultierenden
breiten Tiqqle mit \code{tidyr::fill} aufgefuellt werden (Default:
\emph{FALSE}, d.h. Luecken werden nicht aufgefuellt).}
}
\value{
Tiqqle im breiten Format.
}
\description{
Ein Dataframe wird vom langen ins breite Format umgewandelt. Das lange Format
besitzt hierbei drei Spalten: \emph{time} (Zeitstempel), \emph{signal}
(Signalname) und \emph{value} (Signalwert). Das resultierende breite Format
besitzt eine Spalte \emph{time} (Zeitstempel) und fuer jedes Signal eine
weitere numerische Spalte mit dem Namen des Signals.
}
\seealso{
Other Tiqqle-Funktionen: 
\code{\link{as_long}()},
\code{\link{as_tiqqle}()},
\code{\link{tiqqle}()}
}
\concept{Tiqqle-Funktionen}
