% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{call_fun}
\alias{call_fun}
\title{Dynamischer Funktionsaufruf}
\usage{
call_fun(fun, x, arg1 = NULL, arg2 = NULL, prefix = "")
}
\arguments{
\item{fun}{String mit dem (Stamm-)Namen der aufzurufenden Funktion.}

\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}

\item{arg1}{Benannte Liste mit dem ersten Teil der (ggf. vorgegebenen)
Argumente (Default: \emph{NULL}, d.h. es werden keine weiteren Argumente
beruecksichtigt).}

\item{arg2}{Benannte Liste mit dem zweiten Teil der (ggf. vorberechneten)
Argumente (Default: \emph{NULL}, d.h. es werden keine weiteren Argumente
beruecksichtigt).}

\item{prefix}{String mit Prefix, der dem (Stamm-)Namen beim Funktionsaufruf
vorangestellt wird (Default: \emph{""}, d.h. \code{fun} wird unveraendert
aufgerufen).}
}
\value{
Rueckgabewerte der aufgerufenen Funktion.
}
\description{
Hilfsfunktion zum Aufrufen einer Funktion. Der Funktionsname wird dabei als
Teilstrings uebergeben sowie die Funktionsargumente als einzelne Listen bzw.
als Signal-Dataframe. Aus diesen Bestandteilen wird dann dynamisch ein
Funktionsaufruf erzeugt und ausgefuehrt. Dieser Mechanismus ermoeglicht ein
flexibles Erweitern und Aufrufen von Vorbereitungs-/Extraktionsfunktionen zur
Merkmalsberechnung.
}
\keyword{internal}
