% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{condense}
\alias{condense}
\title{Ohne wiederholte Signalwerte}
\usage{
condense(x)
}
\arguments{
\item{x}{Tiqqle, der reduziert wird.}
}
\value{
Reduzierter Tiqqle \code{x}.
}
\description{
Ueberprueft, ob sich in einem Dataframe aufeinanderfolgende Signalwerte
veraendert haben oder nicht. Falls nicht, werden die redundanten Werte
entfernt. Auf diese Weise wird ein Dataframe mit wiederholungsfreien
Signalwerten erzeugt.
}
\seealso{
\code{\link{value_has_changed}}

Other Reduktionsfunktionen: 
\code{\link{normalize}()},
\code{\link{remove_constant}()}
}
\concept{Reduktionsfunktionen}
