% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{crop}
\alias{crop}
\title{Zeitlich beschnittene Signale}
\usage{
crop(
  x,
  start = NULL,
  end = NULL,
  pad_start = TRUE,
  pad_end = TRUE,
  fill_na = FALSE
)
}
\arguments{
\item{x}{Tiqqle, der geschnitten wird.}

\item{start}{POSIXct-Vektor mit Startzeitpunkten der Fenster (Default:
\emph{NULL}, d.h. der kleinste vorhandene Zeitstempel wird angenommen).}

\item{end}{POSIXct-Vektor mit Endzeitpunkten der Fenster (Default:
\emph{NULL}, d.h. der groesste vorhandene Zeitstempel wird angenommen).}

\item{pad_start}{Logischer Wert, ob geschnittener Dataframe am Anfang mit den
jeweils letzten gueltigen Signalwerten aufgefuellt werden soll (Default:
\emph{TRUE}, d.h. Signale werden aufgefuellt).}

\item{pad_end}{Logischer Wert, ob geschnittener Dataframe am Ende mit den
jeweils letzten gueltigen Signalwerten aufgefuellt werden soll (Default:
\emph{TRUE}, d.h. Signale werden aufgefuellt).}

\item{fill_na}{Logischer Wert, ob beim Auffuellen am Anfang des Dataframes
(Argument \code{pad_start == TRUE}) bei evtl. Auftreten von
\emph{NA}-Werten diese durch den naechsten gueltigen Signalwert ersetzt
werden (Default: \emph{FALSE}, d.h. Signale werden nicht rueckwaerts
aufgefuellt).}
}
\value{
Tiqqle mit zeitlich beschnittenen (und ggf. aufgefuellten)
  Signalen aus \code{x}.
}
\description{
Schneidet aus einem Dataframe mit einer Gesamtsignalmenge eine Reihe von
zeitlichen Fenstern aus. Dabei ist diese Teilmenge durch jeweils einen Start-
und einen Endzeitpunkt definiert. Falls einer dieser beiden Zeitpunkte leer
bleibt, wird automatisch der kleinste bzw. groesste auftretende Zeitstempel
verwendet (Platzhalter).
}
\details{
Um fehlende Signaleintraege am Anfang/Ende des geschnittenen Fensters zu
vermeiden, koennen an diesen Zeitpunkten zusaetzliche Signalwerte eingefuegt
werden. Hierfuer wird der zu diesem Zeitpunkt letzte gueltige Wert bestimmt
und hinzugefuegt (d.h. \emph{Last observation carried forward}). Da u.U. fuer
den Startzeitpunkt des Fensters kein vorheriger Wert existiert, kann in
diesem Fall auch der naechste gueltige Signalwert eingesetzt werden (d.h.
\emph{Next observation carried backward}).
}
