% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{diff_time}
\alias{diff_time}
\title{Zeitdifferenz}
\usage{
diff_time(x, y, units = "secs")
}
\arguments{
\item{x}{POSIXct-Vektor mit Zeitpunkt-Menge \emph{x}.}

\item{y}{POSIXct-Vektor mit Zeitpunkt-Menge \emph{x}.}

\item{units}{String mit der zu verwendenden Zeiteinheit (Default:
\emph{"secs"}, d.h. Sekunden).}
}
\value{
Numerischer Vektor mit Zeitdifferenzen.
}
\description{
Berechnet die Zeitdifferenz zwischen zwei Vektoren in bestimmter Zeiteinheit.
Das Ergebnis wird als numerischer Wert zurueckgegeben.
}
\seealso{
\code{\link{difftime}}
}
\keyword{internal}
