% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{equal}
\alias{equal}
\title{Paarweise identische Ereignisse?}
\usage{
equal(x_start, x_end, y_start, y_end)
}
\arguments{
\item{x_start}{Vektor mit den Startzeitpunkten von Ereignismenge \emph{x}.}

\item{x_end}{Vektor mit den Endzeitpunkten von Ereignismenge \emph{x}.}

\item{y_start}{Vektor mit den Startzeitpunkten von Ereignismenge \emph{y}.}

\item{y_end}{Vektor mit den Endzeitpunkten von Ereignisbenge \emph{y}.}
}
\value{
Logischer Vektor, welche Ereignisse aus \code{x} identisch sind mit
  Ereignis aus \code{y}.
}
\description{
Stellt fuer zwei Mengen von Ereignissen fest, welche paarweise zeitlich
identisch sind. Die Ereignisse werden dabei durch ihre Start-/Endzeitpunkte
beschrieben.
}
\seealso{
Other Testfunktionen: 
\code{\link{include}()},
\code{\link{nearby}()},
\code{\link{overlap}()}
}
\concept{Testfunktionen}
\keyword{internal}
