% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_condense}
\alias{event_condense}
\title{Reduziert ueberlappende Ereignisse}
\usage{
event_condense(x)
}
\arguments{
\item{x}{Dataframe, der die zu reduzierende Menge an Ereignissen beinhaltet.}
}
\value{
Dataframe mit den reduzierten Ereignissen aus \code{x}.
}
\description{
Fasst mehrere sich zeitlich ueberlappende Einzelereignisse zu einem
Gesamtereignis zusammen. Dabei reicht das resultierende Gesamtereignis vom
kleinsten Startzeitpunkt bis zum groessten Endzeitpunkt der Einzelereignisse.
Die Ereignisse werden anhand ihrer Start-/Endzeitpunkte beschrieben, die
jeweils in den Spalten \emph{start} und \emph{end} uebergeben werden.
}
\keyword{internal}
