% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_detect}
\alias{event_detect}
\title{Ereignisse in Signalen}
\usage{
event_detect(x, which, op = `==`, test_value, label = NULL)
}
\arguments{
\item{x}{Tiqqle, der durchsucht werden soll.}

\item{which}{Character-Vektor, der die zu durchsuchenden Signale beinhaltet.}

\item{op}{Funktion, die zum Testen der Signalwerte verwendet wird (Default:
\code{`==`}, d.h. es wird auf Gleichheit getestet - z.B. \code{value ==
test_value}).}

\item{test_value}{Wert, der den Signalwert beschreibt, auf den getestet wird.}

\item{label}{String mit dem Namen des Ereignisses, das ggf. dem Ergebnis als
zusaetzliche Spalte hinzugefuegt wird  (Default: \emph{NULL}, d.h. keine
Namensspalte).}
}
\value{
Dataframe mit den gefundenen Ereignissen.
}
\description{
Sucht in einer Signalmenge nach bestimmten Ereignissen. Dabei ist ein
Ereignis durch den Zeitpunkt des Erfuellens einer bestimmten Bedingungen
(bezueglich des Signalwerts) gekennzeichnet (z.B. \code{value == 1} oder
\code{value != 0}). Sollte die Bedingung an mehreren direkt
aufeinanderfolgenden Zeitpunkten erfuellt sein, wird nur der erste Zeitpunkt
zurueckgegeben.
}
\seealso{
Other Ereignisfunktionen: 
\code{\link{event_match_seq}()},
\code{\link{event_target}()},
\code{\link{event_test}()}
}
\concept{Ereignisfunktionen}
