% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_invert}
\alias{event_invert}
\title{Invertiert Ereignisse}
\usage{
event_invert(x, int_start = NULL, int_end = NULL)
}
\arguments{
\item{x}{Dataframe, der die zu invertierende Menge an Ereignissen beinhaltet.}

\item{int_start}{POSIXct-Zeitstempel mit impliziten Startzeitpunkt des
Invertierungszeitraums (Default: \emph{NULL}, d.h. nur explizite
Ereignisse).}

\item{int_end}{POSIXct-Zeitstempel mit impliziten Endzeitpunkt des
Invertierungszeitraums (Default: \emph{NULL}, d.h. nur explizite
Ereignisse).}
}
\value{
Dataframe mit den invertierten Ereignissen aus \code{x}.
}
\description{
Bestimmt fuer eine Ereignismenge die Differenzmenge. Dabei wird als
Differenz- ereignis der Zeitraum zwischen zwei gegebenen Ereignissen
angenommen. Die Ereignisse werden anhand ihrer Start-/Endzeitpunkte
beschrieben, die jeweils in den Spalten \emph{start} und \emph{end}
uebergeben werden. Zusaetzlich kann noch ein impliziter Invertierungszeitraum
uebergeben werden, mit dem der Anfang und das Ende explizit definiert wird.
Ansonsten entspricht der Anfang dem ersten und das Ende dem letzten
Differenzereignis.
}
\keyword{internal}
