% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_match_freq}
\alias{event_match_freq}
\title{Zueinandergehoerige Ereignisse}
\usage{
event_match_freq(
  x,
  y,
  time_var_x = "time",
  time_var_y = "time",
  tolerance_in_sec = 10
)
}
\arguments{
\item{x}{Dataframe mit der Ereignismenge \emph{x}.}

\item{y}{Dataframe mit der Ereignismenge \emph{y}.}

\item{time_var_x}{Spaltenname (Symbol oder Character) der Zeitvariable in
Dataframe \code{x} (Default: \emph{"time"}).}

\item{time_var_y}{Spaltenname (Symbol oder Character) der Zeitvariable in
Dataframe \code{y} (Default: \emph{"time"}).}

\item{tolerance_in_sec}{Numerischer Wert mit der Laenge des Toleranzfensters
in Sekunden (Default: 10, d.h. Toleranzfenster von 10 Sekunden).}
}
\value{
Dataframe mit den zugeordneten Zeitpunkten.
}
\description{
Ordnet jedem Ereignis aus Menge \code{x} ein passendes Ereignis aus \code{y}
zu. Dabei wird die Zeitdifferenz zwischen den Zeitstempeln der beiden Mengen
betrachtet und die am haeufigsten auftretende Differenz ermittelt. Darauf
aufbauend werden alle Zeitdifferenzen verworfen, die ausserhalb eines
Toleranzfensters um die haeufigste Differenz liegen.
}
\seealso{
\code{\link{diff_time}}
}
\concept{Ereignisgfunktionen}
