% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_merge}
\alias{event_merge}
\title{Verschmilzt angrenzende Ereignisse}
\usage{
event_merge(x, threshold_in_sec = 0)
}
\arguments{
\item{x}{Dataframe, der die zu verschmelzende Menge an Ereignissen
beinhaltet.}

\item{threshold_in_sec}{Numerischer Wert, wie viele Sekunden zwei Ereignisse
maximal auseinander liegen duerfen, um noch als angrenzend zu gelten
(Toleranzbereich) (Default: 0, d.h. sie muessen direkt benachbart sein).}
}
\value{
Dataframe mit den verschmolzenen Ereignissen aus \code{x}.
}
\description{
Fasst mehrere zeitlich angrenzende Einzelereignisse zu einem Gesamtereignis
zusammen. Als angrenzend werden dabei Ereignisse betrachtet, deren zeitlicher
Abstand unterhalb eines Schwellwerts liegt (Toleranzbereich). Das
resultierende Gesamtereignis reicht vom kleinsten Startzeitpunkt bis zum
groessten Endzeitpunkt der Einzelereignisse. Die Ereignisse werden anhand
ihrer Start-/Endzeitpunkte beschrieben, die jeweils in den Spalten
\emph{start} und \emph{end} uebergeben werden.
}
\keyword{internal}
