% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_target}
\alias{event_target}
\title{Klassen fuer Ereignisse}
\usage{
event_target(
  x,
  target_event,
  target_cut_in_sec = NULL,
  add_id = TRUE,
  keep_label = FALSE,
  keep_time = TRUE
)
}
\arguments{
\item{x}{Dataframe, der die zu klassierende Menge an Ereignissen beinhaltet.}

\item{target_event}{Dataframe, der die Zielereignisse beinhaltet. Die
Ereignisse sind anhand ihrer Start-/Endzeitpunkte beschrieben, die jeweils
in den Spalten \emph{start} und \emph{end} uebergeben werden.}

\item{target_cut_in_sec}{Numerischer Wert mit der Laenge des Zeitfensters (in
Sekunden), das vor jedem Zielereignis liegt und den Bereich der potentiell
positiven Ereignisse umfasst (Default: \emph{NULL}, d.h. ist nicht
gesetzt).}

\item{add_id}{Logischer Wert, ob eine ID-Spalte hinzugefuegt werden soll
(Default: \emph{TRUE}).}

\item{keep_label}{Logischer Wert, ob bereits existierende Klassenlabel von
den Zielereignissen uebernommen werden oder ggf. neue berechnet werden
sollen (Default: \emph{FALSE}, d.h. anhand von \code{target_cut_in_sec}
werden neue binaere Label berechnet).}

\item{keep_time}{Logischer Wert, ob Zeitabstand zum naechsten Zielereignis
als Spalte \emph{time_in_sec} beibehalten wird und ggf. als numerische
Regressionsgroesse verwendet werden kann (Default: \emph{TRUE}).}
}
\value{
Dataframe mit den um Klassen ergaenzten Ereignissen aus \code{x}.
}
\description{
Fuegt einer Menge von Ereignissen bestimmte Klasseninformationen hinzu. Diese
wird bezueglich einer Menge an Zielereignissen ermittelt, indem das naechste
nachfolgende Zielereignis bestimmt wird.
}
\seealso{
Other Ereignisfunktionen: 
\code{\link{event_detect}()},
\code{\link{event_match_seq}()},
\code{\link{event_test}()}
}
\concept{Ereignisfunktionen}
