% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_fit_linear}
\alias{extract_fit_linear}
\alias{fit_linear}
\title{Linear-Fit-Merkmale (fit_linear)}
\usage{
extract_fit_linear(x)
}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}
}
\value{
Dataframe mit berechneten Merkmalen.
}
\description{
Berechnet fuer ein Signal eine lineare Regression, welche die Signalwerte
approximiert. Die resultierenden Merkmale heissen \emph{slope} (Anstieg) und
\emph{intercept} (Schnittpunkt y-Achse).
}
\details{
Bei der Uebergabe der Signalfenster an Funktion \code{feature} ist zu
beachten, dass sie eine regulaere Abtastung besitzen sollten, da sonst
verfaelschte Regressionsparameter berechnet werden.
}
\seealso{
Other Merkmalsfunktionen: 
\code{\link{extract_fft_peak}()},
\code{\link{extract_form_bin}()},
\code{\link{extract_raw}()},
\code{\link{extract_stat}()},
\code{\link{prepare_hist_equi}()},
\code{\link{prepare_pca}()}
}
\concept{Merkmalsfunktionen}
