% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_form_bin}
\alias{extract_form_bin}
\alias{form_bin}
\title{Binaere Form-Merkmale (form_bin)}
\usage{
extract_form_bin(x, value_pos = 1, value_neg = 0, is_relative = TRUE)
}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}

\item{value_pos}{Numerischer Wert, der den positiven Signalwert beschreibt
(Default: 1).}

\item{value_neg}{Numerischer Wert, der den negativen Signalwert beschreibt
(Default: 1).}

\item{is_relative}{Logischer Wert, ob Plateau-Merkmale standardisiert werden,
d.h. relativ zur Signalfensterlaenge definiert sind (Default: \emph{TRUE}).}
}
\value{
Dataframe mit berechneten Merkmalen.
}
\description{
Berechnet fuer ein binaeres Signal (d.h. es treten nur zwei verschiedene
Signalwerte auf) die Anzahl von Flanken und die Laengen der Plateaus.
Waehrend eine Flanke hierbei den Uebergang zwischen unterschiedlichen Werten
bezeichnet, besteht ein Plateau fuer die Dauer eines unveraenderlichen Werts.
}
\details{
Die resultierenden Flanken-Merkmale heissen \emph{edge_pos} (Anzahl Flanken
von negativ zu positiv), \emph{edge_neg} (Anzahl Flanken von positiv zu
negativ), \emph{edge_posneg} (Anzahl Flanken von negativ zu positiv und
wieder zurueck), \emph{edge_negpos} (Anzahl Flanken von positiv zu negativ
und wieder zurueck).

Die resultierenden Plateau-Merkmale heissen \emph{plateau_pos} (summierte
Laengen der positiven Plateaus), \emph{plateau_pos} (summierte Laengen der
negativen Plateaus) sowie \emph{plateau_*_mean} (mittlere Laengen der
jeweiligen Plateaus) und \emph{plateau_*_var} (Laengenvarianz der jeweiligen
Plateaus).
}
\seealso{
Other Merkmalsfunktionen: 
\code{\link{extract_fft_peak}()},
\code{\link{extract_fit_linear}()},
\code{\link{extract_raw}()},
\code{\link{extract_stat}()},
\code{\link{prepare_hist_equi}()},
\code{\link{prepare_pca}()}
}
\concept{Merkmalsfunktionen}
