% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{prepare_hist_equi}
\alias{prepare_hist_equi}
\alias{prepare_hist_vary}
\alias{extract_hist}
\alias{extract_hist_equi}
\alias{hist_equi}
\alias{extract_hist_vary}
\alias{hist_vary}
\title{Histogramm-Merkmale (hist_equi, hist_vary)}
\usage{
prepare_hist_equi(x, nr_bins = 1, breaks = NULL, ...)

prepare_hist_vary(x, nr_bins = 1, breaks = NULL, ...)

extract_hist(x, breaks, is_relative = TRUE, ...)

extract_hist_equi(...)

extract_hist_vary(...)
}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}

\item{nr_bins}{Numerischer Wert mit der Anzahl der zu bestimmenden Bins
(Default: 1).}

\item{breaks}{Numerischer Vektor mit den Positionen der Bin-Grenzen, der ggf.
vorberechnet wurde und weitergereicht werden soll (Default: \emph{NULL},
d.h. Breaks werden neu berechnet).}

\item{...}{Weitere Argumente, die durchgereicht werden koennen.}

\item{is_relative}{Logischer Wert, ob eine relative Haeufigkeit berechnet
wird oder eine absolute (Default: \emph{TRUE}, d.h. es wird relativ
gezaehlt).}
}
\value{
Liste mit vorbereiteten Argumenten.

Liste mit vorbereiteten Argumenten.

Dataframe mit berechneten Merkmalen.
}
\description{
Berechnet fuer ein Signal ein Histogramm mit den Haeufigkeiten der
gruppierten Signalwerte (Bins). Dabei werden zur Gruppierung die im Argument
\code{breaks} uebergebenen Bin-Grenzen verwendet.
}
\details{
Waehrend bei der Variante \code{extract_hist_equi} eine bestimmte Anzahl
gleich grosser Bins verwendet wird, verwendet die Variante
\code{extract_hist_vary} unterschiedlich grosse Bins, die auf der Lage der
Quantile beruhen. Zur Bestimmung der jeweiligen Bin-Grenzen werden mittels
der Funktionen \code{prepare_hist_equi} und \code{prepare_hist_vary} die
noetigen Argumente vorberechnet.

Fuer jeden Bin \emph{i} wird das zugehoerigen Merkmal \emph{bin\strong{i}}
berechnet, z.B. \emph{bin1}, \emph{bin2} and \emph{bin3} fuer
\code{length(breaks) == 3}.

Bei der Uebergabe der Signalfenster an Funktion \code{feature} ist zu
beachten, dass sie eine regulaere Abtastung besitzen sollten, da sonst ein
verfaelschtes Histogramm berechnet wird.
}
\section{Functions}{
\itemize{
\item \code{prepare_hist_equi}: Bereitet Histogramme mit festen Bins vor.

\item \code{prepare_hist_vary}: Bereitet Histogramme mit flexiblen Bins vor.

\item \code{extract_hist_equi}: Berechnet Histogramme mit festen Bins.

\item \code{extract_hist_vary}: Berechnet Histogramme mit flexiblen Bins.
}}

\seealso{
Other Merkmalsfunktionen: 
\code{\link{extract_fft_peak}()},
\code{\link{extract_fit_linear}()},
\code{\link{extract_form_bin}()},
\code{\link{extract_raw}()},
\code{\link{extract_stat}()},
\code{\link{prepare_pca}()}
}
\concept{Merkmalsfunktionen}
