% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{extract_raw}
\alias{extract_raw}
\alias{raw}
\title{Roh-Merkmale (raw)}
\usage{
extract_raw(x)
}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}
}
\value{
Dataframe mit berechneten Merkmalen.
}
\description{
Gibt fuer ein Signal unveraendert die Signalwerte zurueck, z.b. die Merkmale
\emph{value1}, \emph{value2} und \emph{value3} fuer eine Signal mit
\code{nrow(x) == 3}.
}
\seealso{
Other Merkmalsfunktionen: 
\code{\link{extract_fft_peak}()},
\code{\link{extract_fit_linear}()},
\code{\link{extract_form_bin}()},
\code{\link{extract_stat}()},
\code{\link{prepare_hist_equi}()},
\code{\link{prepare_pca}()}
}
\concept{Merkmalsfunktionen}
