% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_stat}
\alias{extract_stat}
\alias{stat}
\title{Statistik-Merkmale (stat)}
\usage{
extract_stat(x)
}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}
}
\value{
Dataframe mit berechneten Merkmalen.
}
\description{
Berechnet fuer ein Signal die folgendenen statistischen Kennzahlen:
\emph{min} (Minimum), \emph{max} (Maximum), \emph{mean} (Mittelwert),
\emph{median} (Median) und \emph{var} (Varianz).
}
\details{
Bei der Uebergabe der Signalfenster an Funktion \code{feature} ist zu
beachten, dass sie eine regulaere Abtastung besitzen sollten, da sonst
verfaelschte statistsche Kennzahlen berechnet werden.
}
\seealso{
Other Merkmalsfunktionen: 
\code{\link{extract_fft_peak}()},
\code{\link{extract_fit_linear}()},
\code{\link{extract_form_bin}()},
\code{\link{extract_raw}()},
\code{\link{prepare_hist_equi}()},
\code{\link{prepare_pca}()}
}
\concept{Merkmalsfunktionen}
