% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature}
\alias{feature}
\title{Merkmale fuer Fenster}
\usage{
feature(
  x,
  feat,
  arg_global = TRUE,
  as_df = TRUE,
  as_wide = TRUE,
  .progress = TRUE,
  ...
)
}
\arguments{
\item{x}{Liste mit Fenster-Tiqqles (bzw. ein Einzel-Tiqqle), fuer die
Merkmale berechnet werden.}

\item{feat}{Benannte Liste mit den zu berechnenden Merkmalen und den
zugehoerigen Argumenten. Siehe Details fuer ausfuehrlichere Erlaeuterungen.}

\item{arg_global}{Logischer Wert, ob evtl. benoetigte Funktionsargumente
global ueber alle Fenster vorberechnet werden oder lokal fuer die
Einzelfenster (Default: \emph{TRUE}, d.h. sie werden global berechnet).}

\item{as_df}{Logischer Wert, ob Merkmale als Dataframe zurueckgegeben werden
oder als Liste (Default: \emph{TRUE}, d.h. Merkmals-Dataframe wird
erzeugt).}

\item{as_wide}{Logischer Wert, ob Merkmals-Dataframe ggf. im breiten Format
zurueckgegeben wird oder im langen Format (Default: \emph{TRUE}, d.h.
Merkmals-Dataframe wird im breiten Format erzeugt).}

\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}

\item{...}{Weitere Argumente, die an Funktion \code{feature_df} durchgereicht
werden.}
}
\value{
Liste oder Dataframe mit berechneten Merkmalen.
}
\description{
Fuer eine Menge von Fenstern werden Merkmalsbeschreibungen extrahiert.
Hierbei wird zunaechst ein Vorbereitungs- und anschließend ein
Extraktionsschritt durchgefuehrt. Die resultierenden Merkmale stellen eine
kompakte Repraesentation des Signalverlaufs dar und koennen als
Trainingsdaten fuer maschinelles Lernen verwendet werden.
}
\details{
Die zu berechnenden Merkmale werden mittels des Arguments \code{feat}
ausgewaehlt. Dieses erwartet eine benannte Liste mit je einem Eintrag pro
Merkmalstyp. Waehrend der Name des Listeneintrags den Namen des zu
berechnenden Merkmalstyp angibt (und somit die aufzurufenden Funktionen),
beschreibt das Listenelement die hierfuer zu verwendenden Funktionsargumente.
Ein Listenelement besteht dabei wiederum aus einer benannten Liste. Der Namen
des Eintrags beschreibt den zu setzenden Argumentnamen, der Wert des Eintrags
den entsprechenden zu uebergebenden Argumentwert. Falls kein zusaetzlicher
Argumentwert existiert oder die Default-Werte verwendet werden sollen, wird
hier eine leere Liste \code{list()} oder der Wert \emph{NULL} uebergeben.

Die verfuegbaren Merkmalstypen koennen mit der Hilfsfunktion
\code{feature_all} aufgelistet werden. Aktuell sind dies:
\itemize{
  \item \link{fft_peak}
  \item \link{fit_linear}
  \item \link{form_bin}
  \item \link{hist_equi}
  \item \link{hist_vary}
  \item \link{raw}
  \item \link{pca}
  \item \link{stat}
}

In den jeweiligen Funktionen sind die verfuegbaren Argumente dokumentiert.
Fuer einige Merkmalsberechnungen ist es noetig, im Vorbereitungsschritt
gewisse Argumente global vorzuberechnen. Dieses Verhalten kann durch das
Argument \code{arg_global} gesteuert werden.

Das Ergebnis der Merkmalsberechnung liegt intern als geschachtelte Liste vor.
Diese Merkmalsliste beinhaltet fuer jedes Fenster einen Eintrag mit einem
Dataframe, der fuer jedes Signal wiederum eine benannte Liste mit den
berechneten Merkmale umfasst. Diese Liste kann mittels der Argumente
\code{as_df} und \code{as_wide} in einen Dataframe im passenden Format
konvertiert werden.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

\seealso{
\code{\link{feature_all}}, code{\link{feature_df}},
  \code{\link{feature_wide}}
}
