% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_all}
\alias{feature_all}
\title{Alle Merkmalstypen}
\usage{
feature_all()
}
\value{
Character-Vektor mit den verfuegbaren Merkmalstypen.
}
\description{
Listet alle verfuegbaren Merkmalstypen auf, die als potentielle Argumente an
Funktion \code{feature} uebergeben werden koennen.
}
