% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_context}
\alias{feature_context}
\title{Merkmale in Kontextfenstern}
\usage{
feature_context(
  x,
  window_length,
  window_step,
  prefix = "context",
  .progress = FALSE
)
}
\arguments{
\item{x}{Dataframe mit Merkmalen im breiten Format.}

\item{window_length}{Numerischer Wert mit der Fensterlaenge, d.h. Anzahl der
Kontextfenster, die zusammen einen Kontext bilden.}

\item{window_step}{Numerischer Wert mit der Schrittweite, um welche die
Fenster zeitlich weitergeschoben werden.}

\item{prefix}{String mit Prefix, der den neuen Merkmalsnamen (Spalten)
vorangestellt wird (Default: \emph{"context"}).}

\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}
}
\value{
Dataframe mit Merkmalen in Kontextfenstern.
}
\description{
Stellt fuer eine Menge von Merkmalen einen zeitlichen Verlauf her, der als
Kontext bezeichnet wird. Ein einzelnes Kontextfenster umfasst hierbei die mit
der Funktion \code{feature} fuer ein Signalfenster berechneten Merkmale und
entspricht in dem resultierenden breiten Merkmals-Dataframe einer Zeile. Der
Gesamtkontext umfasst hierbei die Menge der gleitenden Kontextfenster, deren
Zusammensetzung und Anordnung mittels \code{window_length} (Fensterlaenge)
und \code{window_step} (Schrittweite) angegeben werden.
}
\details{
Auf diese Weise werden mehrere Kontextfenster in einer Zeile des neuen
Merkmals-Dataframe zusammengefasst. Durch die Benennung der auf diese Weise
entstehenden Spalten koennen Rueckschluesse auf die urspruengliche Benennung
gezogen werden. Das im aktuellen Kontext zeitlich gesehen letzte
Kontextfenster hat hierbei den Index 0, das vorherige Fenster entsprechend
den Index 1 usw. Falls der Ausgangs-Dataframe eine Spalte mit dem Namen
\emph{id} besitzt, so wird diese beibehalten und angepasst.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

