% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_df}
\alias{feature_df}
\title{Konvertiert Merkmalsliste in Dataframe}
\usage{
feature_df(x, add_id = TRUE)
}
\arguments{
\item{x}{Liste mit den berechneten Merkmalen.}

\item{add_id}{Logischer Wert, ob eine ID-Spalte hinzugefuegt werden soll
(Default: \emph{TRUE}).}
}
\value{
Merkmals-Dataframe im langen Format.
}
\description{
Eine Merkmalsliste, wie sie von der Funktion \code{feature} erzeugt wird,
wird in einen Dataframe im langen Format umgewandelt.
}
\details{
Der resultierende Dataframe umfasst die Spalten \emph{feature_set}
(Merkmalsmenge), \emph{feature} (Merkmalsnamen), \emph{signal} (Signalnamen
fuer den das Merkmal berechnet wurde) sowie \emph{value} (Merkmalswert).
Zusaetzlich kann eine Spalte \emph{id} hinzugefuegt werden, welche die
Fenster-ID beinhaltet.

Mit der ergaenzenden Funktion \code{feature_wide} kann der Merkmals-Dataframe
in ein breites Format konvertiert werden.
}
\seealso{
\code{\link{feature}}, \code{\link{feature_wide}}
}
\keyword{internal}
