% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_dummy}
\alias{feature_dummy}
\title{Doku-Dummyfunktion}
\arguments{
\item{x}{Dataframe, der das zu verarbeitetende Signal enthaelt und mindestens
eine Spalte \emph{value} besitzt.}

\item{...}{Weitere Argumente, die durchgereicht werden koennen.}
}
\description{
Hilfsfunktion zu Dokumentationszwecken.
}
\keyword{internal}
