% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_wide}
\alias{feature_wide}
\title{Konvertiert langen Merkmals-Dataframe in breites Format}
\usage{
feature_wide(x)
}
\arguments{
\item{x}{Merkmals-Dataframe im langen Format, der konvertiert wird.}
}
\value{
Merkmals-Dataframe im breiten Format.
}
\description{
Ein Merkmals-Dataframe wird vom langen ins breite Format umgewandelt.
}
\details{
Das lange Format, wie es von der Funktion \code{feature_df} erzeugt wird,
besitzt hierbei die Spalten \emph{feature_set} (Merkmalsmenge),
\emph{feature} (Merkmalsnamen), \emph{signal} (Signalnamen fuer den das
Merkmal berechnet wurde) sowie \emph{value} (Merkmalswert). Zusaetzlich kann
die Spalte \emph{id} mit der Fenster-ID vorhanden sein. Das resultierende
breite Format besitzt fuer jede Kombination aus Merkmalsmenge/-name und
Signal eine entsprechende Spalte mit dem berechneten Merkmalswert. Diese
Spalten sind nach dem Schema \emph{[feature_set]-[feature]-[signal]} benannt.
Falls vorhanden, wird die Spalte \emph{id} uebernommen.
}
\seealso{
\code{\link{feature}}, \\code{\link{feature_df}}
}
\keyword{internal}
