% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{fill_signal_wide}
\alias{fill_signal_wide}
\title{Letzter gueltiger Signalwert im breiten Format}
\usage{
fill_signal_wide(x, fill_na = TRUE, direction = "forward")
}
\arguments{
\item{x}{Tiqqle, aus dem der gueltige Wert bestimmt wird.}

\item{fill_na}{Logischer Wert, ob evtl. entstehende \emph{NA}-Werte mit dem
naechsten gueltigen Signalwert ersetzt werden (Default: \emph{TRUE}, d.h.
sie werden ersetzt).}

\item{direction}{String mit zeitlicher Ersetzungsrichtung - entweder
\emph{"forward"} (Default: letzter gueltiger Wert ) oder \emph{"backward"}
(naechster gueltiger Wert).}
}
\value{
Tiqqle mit letzten gueltigen Signalwerten aus \code{x}.
}
\description{
Hilfsfunktion zur Bestimmung des letzten oder naechsten Signalwerts im
breiten Format.
}
\seealso{
\code{\link{last_value.tiqqle_wide}}
}
\keyword{internal}
