% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{first_row_per}
\alias{first_row_per}
\title{Erste Zeile pro Gruppe aus Dataframe}
\usage{
first_row_per(x, var)
}
\arguments{
\item{x}{Dataframe, der gefiltert wird und bereits in der gewuenschten
Sortierung vorliegt. Kann ausserdem bereits gruppiert sein.}

\item{var}{Spaltenname (Symbol) der gruppierenden Variable.}
}
\value{
Dataframe mit den jeweils ersten Zeilen pro Gruppe aus \code{x}.
}
\description{
Filtert aus einem Dataframe die jeweils gruppenweise erste Zeile heraus.
Dabei wird die urspruengliche (zeitliche) Sortierung beruecksichtigt, d.h.
pro sequenziellem Auftreten der Gruppenvariable wird eine Zeile ausgewaehlt.
Dieses Verhalten unterscheidet sich von \code{x \%>\% group_by(var) \%>\%
slice(1)}.
}
\details{
Falls der Dataframe bereits nach einem zusaetzlichen Kriterium gruppiert ist
(z.B. nach Signal), bleibt die bestehende Gruppierung erhalten und wird um
die lokale Gruppierung erweitert.
}
\keyword{internal}
