% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{future_dummy}
\alias{future_dummy}
\title{Doku-Dummyfunktion}
\arguments{
\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}
}
\description{
Hilfsfunktion zu Dokumentationszwecken.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

\keyword{internal}
