% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{harmonize}
\alias{harmonize}
\title{Harmonisierter Dataframe}
\usage{
harmonize(
  x,
  time_var = "time",
  signal_var = "signal",
  value_var = "value",
  format = "long",
  ...
)
}
\arguments{
\item{x}{Dataframe, der in die standardisierte Form gebracht werden soll.}

\item{time_var}{Spaltenname (Symbol oder String) der Zeitvariable im
Dataframe \code{x} (Default: \emph{"time"}).}

\item{signal_var}{Spaltenname (Symbol oder String) der
Signalnamen-Variable im Dataframe \code{x} (Default: \emph{"signal"}).}

\item{value_var}{Spaltenname (Symbol oder String) der Signalwert-Variable
im Dataframe \code{x} (Default: \emph{"value"}).}

\item{format}{String, welches Format der erzeugte Tiqqle hat. Erlaubte Werte:
\emph{"long"} oder \emph{"wide"} (Default: \emph{"long"}, d.h. langes
Format).}

\item{...}{Weitere Argumente, die ggf. an Funktion \code{as_wide}
durchgereicht werden.}
}
\value{
Tiqqle im gewuenschten Format.
}
\description{
Ueberfuehrt einen Signal-Dataframe in einen standardisierten Tiqqle. Dabei
wird angenommen, dass der urspruengliche Dataframe ein langes Format besitzt.
}
\seealso{
\code{\link{as_wide}}

Other Harmonisierungsfunktionen: 
\code{\link{harmonize_event}()}
}
\concept{Harmonisierungsfunktionen}
