% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{harmonize_event}
\alias{harmonize_event}
\title{Harmonisierter Ereignis-Dataframe}
\usage{
harmonize_event(x, start_var = "start", end_var = "end", ...)
}
\arguments{
\item{x}{Dataframe, der in die standardisierte Form gebracht werden soll.}

\item{start_var}{Spaltenname (Symbol oder String) der Variable mit dem
Zeitpunkt des Eintretens des Ereignisses im Dataframe \code{x} (Default:
\emph{"start"}).}

\item{end_var}{Spaltenname (Symbol oder String) der Variabele mit dem
Zeitpunkt des Aufhebens des Ereignisses im Dataframe \code{x} (Default:
\emph{"end"}).}

\item{...}{Weitere Spaltennamen (Symbole oder Character), die als
Zusatzinformationen zu den Ereignissen als Faktoren beibehalten werden.}
}
\value{
Dataframe mit den Ereignissen.
}
\description{
Ueberfuehrt einen Ereignis-Dataframe in ein standardisiertes Format.
}
\seealso{
Other Harmonisierungsfunktionen: 
\code{\link{harmonize}()}
}
\concept{Harmonisierungsfunktionen}
