% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_constant}
\alias{is_constant}
\title{Konstanter Vektor?}
\usage{
is_constant(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vektor, der ueberprueft wird.}

\item{na.rm}{Logischer Wert, ob vorhandene \emph{NA}-Werte ignoriert werden
(Default: \emph{TRUE}, d.h. sie werden nicht beruecksichtigt).}
}
\value{
Logischer Wert, ob Vektor \code{x} konstant ist.
}
\description{
Ueberprueft, ob in einem Vektor nur ein einziger Wert vorkommt.
}
\keyword{internal}
