% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_long}
\alias{is_long}
\title{Langer Dataframe?}
\usage{
is_long(x)
}
\arguments{
\item{x}{Dataframe, dessen Format ueberprueft wird.}
}
\value{
Logischer Wert, ob Dataframe \code{x} im langen Format vorliegt.
}
\description{
Ueberprueft, ob ein Dataframe im langen Format vorliegt. Das lange Format
besitzt hierbei drei Spalten: \emph{time} (Zeitstempel), \emph{signal}
(Signalname als Faktor oder String) und \emph{value} (numerischer
Signalwert).
}
\seealso{
Other Format-Checks: 
\code{\link{is_tiqqle}()},
\code{\link{is_valid}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
