% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_null_or}
\alias{is_null_or}
\alias{\%is_null_or\%}
\title{Erfuellt Vektor erweiterte Bedingung?}
\usage{
is_null_or(x, condition, ...)

x \%is_null_or\% condition
}
\arguments{
\item{x}{Vektor, der ueberprueft wird.}

\item{condition}{Funktion, die zum erweiterten Testen von \code{x} verwendet
wird.}

\item{...}{Weitere Argumente, die an Funktion \code{condition} durchgereicht
werden.}
}
\value{
Logischer Wert, ob \code{x} NULL ist oder die Bedingung erfuellt.
}
\description{
Ueberprueft, ob ein Vektor entweder NULL ist oder eine zusaetzliche Bedingung
erfuellt.
}
\keyword{internal}
