% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_valid}
\alias{is_valid}
\title{Valider Tiqqle?}
\usage{
is_valid(x, is_empty = FALSE)
}
\arguments{
\item{x}{Dataframe (\code{\link{tibble}}), dessen Format ueberprueft wird.}

\item{is_empty}{Logischer Wert, ob Dataframe auch leer sein darf, um als
valide zu gelten (Default: \emph{FALSE}, d.h. er muss Daten enthalten).}
}
\value{
Logischer Wert, ob Dataframe \code{x} ein valider Tiqqle ist.
}
\description{
Ueberprueft, ob ein Dataframe ein langer oder breiter Tiqqle ist und ggf.
Daten enthaelt.
}
\seealso{
Other Format-Checks: 
\code{\link{all_are_long}()},
\code{\link{all_are_valid}()},
\code{\link{all_are_wide}()},
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
