% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_wide}
\alias{is_wide}
\title{Breiter Dataframe?}
\usage{
is_wide(x)
}
\arguments{
\item{x}{Dataframe, dessen Format ueberprueft wird.}
}
\value{
Logischer Wert, ob Dataframe \code{x} im breiten Format vorliegt.
}
\description{
Ueberprueft, ob ein Dataframe im breiten Format vorliegt. Das breite Format
besitzt hierbei eine Spalte \emph{time} (Zeitstempel) und fuer jedes Signal
eine weitere numerische Spalte mit dem Namen des Signals.
}
\seealso{
Other Format-Checks: 
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_valid}()}
}
\concept{Format-Checks}
