% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{last_value}
\alias{last_value}
\alias{last_value.tiqqle_long}
\alias{last_value.tiqqle_wide}
\title{Letzter gueltiger Signalwert}
\usage{
last_value(x, before, fill_na = FALSE)

\method{last_value}{tiqqle_long}(x, before, fill_na = FALSE)

\method{last_value}{tiqqle_wide}(x, before, fill_na = FALSE)
}
\arguments{
\item{x}{Tiqqle, aus dem der letzte gueltige Wert bestimmt wird.}

\item{before}{POSIXct-Zeitstempel vor dem der gueltige Wert bestimmt wird.}

\item{fill_na}{Logischer Wert, ob evtl. entstehende \emph{NA}-Werte mit dem
naechsten gueltigen Signalwert ersetzt werden (Default: \emph{FALSE}, d.h.
sie werden nicht ersetzt).}
}
\value{
Tiqqle mit letzten gueltigen Signalwerten aus \code{x}.
}
\description{
Bestimmt den letzten Wert pro Signal, der vor einem bestimmten Zeitpunkt
gueltig war. Falls dieser nicht gefunden wird (d.h. resultierendes
\emph{NA}), kann der naechste nach diesem Zeitpunkt gueltige Wert
herangezogen werden.
}
\section{Methods (by class)}{
\itemize{
\item \code{tiqqle_long}: Letzter gueltiger Wert im langen Format

\item \code{tiqqle_wide}: Letzten gueltiger Wert im breiten Format
}}

\keyword{internal}
