% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{limit}
\alias{limit}
\title{Signalwert-Bereinigung}
\usage{
limit(x, which, when = NULL, op = `-`, value_before, value_after = NULL)
}
\arguments{
\item{x}{Tiqqle, der korrigiert wird.}

\item{which}{Character-Vektor, der das anzupassende Signal beinhaltet.}

\item{when}{POSIXct-Zeitstempel mit Zeitpunkt des Schwellwertwechsels
(Default: \emph{NULL}, d.h. ein Schwellwertwechsel findet nicht statt).}

\item{op}{Funktion, die zum Anpassen der Signalwerte verwendet wird (Default:
\code{`-`}, d.h. Schwellwert wird vom Signalwert subtrahiert).}

\item{value_before}{Numerischer Wert mit dem vor Zeitpunkt \code{when}
gueltigen Schwellwert.}

\item{value_after}{Numerischer Wert mit dem ab Zeitpunkt \code{when}
gueltigen Schwellwert (Default: \emph{NULL}, d.h. \code{value_after}
identisch zu \code{value_before}).}
}
\value{
Tiqqle mit korrigiertem Signalwert.
}
\description{
Korrigiert Signalwert um einen bestimmten Offset, der zu einem gewissen
Zeitpunkt wechseln kann.
}
\concept{Transformationsfunktionen}
