% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{lvls_subset}
\alias{lvls_subset}
\title{Selektiert Faktor-Level}
\usage{
lvls_subset(x, pattern = ".*")
}
\arguments{
\item{x}{Faktor, dessen Level ausgewaehlt werden.}

\item{pattern}{String mit regulaerem Ausdruck zur Selektion der Level
(Default: \emph{".*"}, d.h. alle Level werden ausgewaehlt).}
}
\value{
Character-Vektor mit ausgewaehlten Levels aus \code{x}.
}
\description{
Waehlt eine Teilmenge der verfuegbaren Level eines Faktors aus. Die Selektion
wird hierbei mit einem regulaeren Ausdruck beschrieben.
}
\keyword{internal}
