% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{name_seq}
\alias{name_seq}
\title{Durchnummerierte Namen}
\usage{
name_seq(n, prefix = "", suffix = "")
}
\arguments{
\item{n}{Numerischer Wert mit der Anzahl der zu erzeugenden Namen.}

\item{prefix}{String der den Namen vorangestellt wird (Default: \emph{""},
d.h. vorher wird nichts angefuegt).}

\item{suffix}{String der den Namen hintenangestellt wird (Default: \emph{""},
d.h. hinten wird nichts angefuegt).}
}
\value{
Character-Vektor der durchnummerierten Namen.
}
\description{
Erzeugt \code{n} durchnummerierte Namen, die mit einem Praefix und einem
Suffix versehen sein koennen.
}
\keyword{internal}
