% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{name_unique}
\alias{name_unique}
\title{Eindeutige Namen}
\usage{
name_unique(x)
}
\arguments{
\item{x}{Character-Vektor mit den Ausgangsnamen.}
}
\value{
Character-Vektor der eindeutigen Namen.
}
\description{
Erzeugt fuer eine Ausgangsmenge von Namen \code{x} eine zugehoerige Zielmenge
mit eindeutigen Namen. Einmalig auftretende Namen werden dabei unveraendert
uebernommen. Mehrfach auftretende Namen werden durch Anhaengen einer
fortlaufenden Nummer eindeutig gemacht.
}
\keyword{internal}
