% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{norm_range}
\alias{norm_range}
\title{Normiert Vektor (Min-Max)}
\usage{
norm_range(x, min = 0, max = 1)
}
\arguments{
\item{x}{Vektor, der normiert wird. Kann logisch oder numerisch sein.}

\item{min}{Numerischer Wert, der das neue Minimum angibt (Default: 0).}

\item{max}{Numerischer Wert, der das neue Maximum angibt (Default: 1).}
}
\value{
Normierter Vektor \code{x}.
}
\description{
Bildet den Wertebereich eines Vektors in einen neuen Wertebereich ab. Dabei
wird eine lineare Abbildung vorgenommen (evtl. vorhandene \emph{NA}-Werte
werden beibehalten). Das Mininum und Maximum des neuen Wertebereichs koennen
per Argument angepasst werden. Standardmaessig entspricht dies einer
0-1-Normierung.
}
\keyword{internal}
