% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{normalize}
\alias{normalize}
\title{Signalnormalisierung}
\usage{
normalize(x, fun = norm_range, ...)
}
\arguments{
\item{x}{Tiqqle, der normalisiert wird.}

\item{fun}{Funktion, die zur Normalisierung der Signalwerte verwendet wird
(Default: \code{\link{norm_range}}).}

\item{...}{Weitere Argumente, die an Funktion \code{fun} durchgereicht
werden.}
}
\value{
Normalisierter Tiqqle \code{x}.
}
\description{
Normalisiert die Signale eines Dataframes, indem deren Werte in einen
einheitlichen Wertebereich ueberfuehrt werden. Standardmaessig wird eine
0-1-Normierung angewandt.
}
\details{
Prinzipiell laesst sich hierfuer auch eine andere Normalisierungsfunktion
einsetzen. Diese vorher zu definierende Funktion bekommt als erstes Argument
den numerischen Vektor der Signalwerte uebergeben und gibt einen entsprechend
normierten Vektor zurueck.
}
\seealso{
\code{\link{norm_range}}

Other Reduktionsfunktionen: 
\code{\link{condense}()},
\code{\link{remove_constant}()}
}
\concept{Reduktionsfunktionen}
