% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{overlap}
\alias{overlap}
\title{Paarweise ueberlappende Ereignisse?}
\usage{
overlap(x_start, x_end, y_start, y_end, is_strict = TRUE)
}
\arguments{
\item{x_start}{Vektor mit den Startzeitpunkten von Ereignismenge \emph{x}.}

\item{x_end}{Vektor mit den Endzeitpunkten von Ereignismenge \emph{x}.}

\item{y_start}{Vektor mit den Startzeitpunkten von Ereignismenge \emph{y}.}

\item{y_end}{Vektor mit den Endzeitpunkten von Ereignismenge \emph{y}.}

\item{is_strict}{Logischer Wert, ob eine minimale zeitliche Schnittmenge
zwischen den Ereignissen vorliegen muss, um als einander ueberlappend zu
gelten. Anderenfalls gelten direkt angrenzende Ereignisse ebenfalls als
ueberlappend (Default: \emph{TRUE}, d.h. die strenge Definition gilt).}
}
\value{
Logischer Vektor, welche Ereignisse aus \code{x} sich mit
  irgendwelchen Ereignissen aus \code{y} ueberlappen.
}
\description{
Stellt fuer zwei Mengen von Ereignissen fest, welche sich paarweise zeitlich
ueberlappen. Die Ereignisse werden dabei durch ihre Start-/Endzeitpunkte
beschrieben.
}
\seealso{
Other Testfunktionen: 
\code{\link{equal}()},
\code{\link{include}()},
\code{\link{nearby}()}
}
\concept{Testfunktionen}
\keyword{internal}
