% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_checkered}
\alias{plot_checkered}
\title{Schachbrett-Signalplot}
\usage{
plot_checkered(x, drop_level = TRUE, has_legend = FALSE)
}
\arguments{
\item{x}{Tiqqle im langen Format.}

\item{drop_level}{Logischer Wert, ob nicht vorhandene Signale ausgeblendet
werden (Default: \emph{TRUE}).}

\item{has_legend}{Logischer Wert, ob eine Legende angezeigt wird (Default:
\emph{FALSE}, d.h. keine Legende).}
}
\value{
Plot-Objekt (\code{ggplot2}) mit der Schachbrett-Darstellung.
}
\description{
Plottet Signale als Schachbrett. Dabei verlaeuft die Zeit auf der X-Achse und
die Signale auf der Y-Achse. An den entsprechenden X-Y-Kreuzungspunkten
werden die Signalwerte als farbige Rechtecke angezeigt. Diese entspricht
einer Heatmap-Darstellung.
}
\details{
Die Reihenfolge der Signale (Y-Achse) laesst sich ueber die Level der
Signalspalte festlegen.
}
\seealso{
Other Plotfunktionen: 
\code{\link{plot_spread}()},
\code{\link{plot_stacked}()}
}
\concept{Plotfunktionen}
