% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dummy}
\alias{plot_dummy}
\title{Doku-Dummyfunktion}
\arguments{
\item{x}{Tiqqle im langen Format.}

\item{drop_level}{Logischer Wert, ob nicht vorhandene Signale ausgeblendet
werden (Default: \emph{TRUE}).}

\item{has_legend}{Logischer Wert, ob eine Legende angezeigt wird (Default:
\emph{FALSE}, d.h. keine Legende).}
}
\description{
Hilfsfunktion zu Dokumentationszwecken.
}
\keyword{internal}
