% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spread}
\alias{plot_spread}
\title{Facet-Signalplot}
\usage{
plot_spread(x, nrow = NULL, ncol = 1, drop_level = TRUE)
}
\arguments{
\item{x}{Tiqqle im langen Format.}

\item{nrow}{Numerischer Wert mit der Anzahl der Facet-Rasterzeilen (Default:
\emph{NULL}, d.h. wird automatisch angepasst).}

\item{ncol}{Numerischer Wert mit der Anzahl der Facet-Rasterspalten (Default:
1, d.h. Raster mit einer Spalte).}

\item{drop_level}{Logischer Wert, ob nicht vorhandene Signale ausgeblendet
werden (Default: \emph{TRUE}).}
}
\value{
Plot-Objekt (\code{ggplot2}) mit der Facet-Zeitreihendarstellung.
}
\description{
Plottet Signale als Liniengraphen in unterschiedliche Facets. Innerhalb einer
Facet ist die Zeit auf der X-Achse und die Signalwerte auf der Y-Achse
dargstellt. Die Facets werden dabei in einem bestimmten Raster angeordnet.
}
\details{
Die Reihenfolge der Signale (Facets) laesst sich ueber die Level der
Signalspalte festlegen.
}
\seealso{
Other Plotfunktionen: 
\code{\link{plot_checkered}()},
\code{\link{plot_stacked}()}
}
\concept{Plotfunktionen}
