% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stacked}
\alias{plot_stacked}
\title{Signalplot}
\usage{
plot_stacked(x, drop_level = TRUE, has_legend = FALSE)
}
\arguments{
\item{x}{Tiqqle im langen Format.}

\item{drop_level}{Logischer Wert, ob nicht vorhandene Signale ausgeblendet
werden (Default: \emph{TRUE}).}

\item{has_legend}{Logischer Wert, ob eine Legende angezeigt wird (Default:
\emph{FALSE}, d.h. keine Legende).}
}
\value{
Plot-Objekt (\code{ggplot2}) mit der ueberlagerten
  Zeitreihendarstellung.
}
\description{
Plottet Signale als Liniengraphen ueberlagert in einem Facet. Dabei ist die
Zeit auf der X-Achse und die Signalwerte auf der Y-Achse dargstellt. Die
Signale werden dabei farbig kodiert.
}
\details{
Die Reihenfolge der Signale (Farben) laesst sich ueber die Level der
Signalspalte festlegen.
}
\seealso{
Other Plotfunktionen: 
\code{\link{plot_checkered}()},
\code{\link{plot_spread}()}
}
\concept{Plotfunktionen}
