% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{random_event}
\alias{random_event}
\title{Zufaelliges Ereignisse}
\usage{
random_event(int_start, int_end, offtime, event_length_in_sec, .max_run = 1000)
}
\arguments{
\item{int_start}{POSIXct-Zeitstempel mit Startzeitpunkt des Sample-Zeitraums.}

\item{int_end}{POSIXct-Zeitstempel mit Endzeitpunkt des Sample-Zeitraums.}

\item{offtime}{Dataframe, der die Offtime-Ereignisse beinhaltet. Die
Ereignisse sind anhand ihrer Start-/Endzeitpunkte beschrieben, die jeweils
in den Spalten \emph{start} und \emph{end} uebergeben werden.}

\item{event_length_in_sec}{Numerischer Wert mit der Laenge der Ereignisse (in
Sekunden).}

\item{.max_run}{Numerischer Wert mit Anzahl der Versuche, die je Ereignis
unternommen werden sollen, um es im Zeitraum zu platzieren (Default: 1E3).}
}
\value{
Dataframe mit zufaelligem Ereignis im Zeitraum bzw. \emph{NULL}, wenn
  Platzierung nicht erfolgreich.
}
\description{
Hilfsfunktion zum zufaelligen Platzieren eines Ereignisses innerhalb eines
Zeitraums. Der Gesamtzeitraum wird durch einen Start- und einen Endzeitpunkt
festgelegt. In diesem Zeitraum koennen zusaetzliche Offtime-Intervalle
definiert werden, fuer die beim Platzieren sichergestellt wird, dass keine
Ueberlappung auftritt. D.h. sollte ein Ueberlappung auftreten, wird das
Ereignis verworfen und ein neuer Platzierungsversuch gestartet. Um an dieser
Stelle einen Deadlock zu vermeiden, gibt es eine Obergrenze fuer die
durchzufuehrenden Versuche.
}
\seealso{
\code{\link{sample_random}}

Other Sampling-Funktionen: 
\code{\link{sample_balanced}()},
\code{\link{sample_random}()}
}
\concept{Sampling-Funktionen}
\keyword{internal}
