% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{regularize}
\alias{regularize}
\title{Regularisierte zeitliche Signalaufloesung}
\usage{
regularize(x, insert_missing = TRUE, fill_gap = FALSE, by = "sec")
}
\arguments{
\item{x}{Tiqqle, der regularisiert wird.}

\item{insert_missing}{Logischer Wert, ob auch Zeitstempel und Signalwerte
eingefuegt werden, fuer die kein expliziter regulaerer Wert vorliegt
(Default: \emph{TRUE}, d.h. alle Zeitstempel werden aufgefuellt).}

\item{fill_gap}{Logischer Wert, ob entstehende \emph{NA}-Werte mittels
\code{\link{fill}} aufgefuellt werden (Default: \emph{FALSE}, d.h. sie
werden nicht veraendert).}

\item{by}{String mit dem zu verwendenden Zeitintervall (Default:
\emph{"sec"}, d.h. Sekundenintervall). Siehe \code{\link{seq.POSIXt}} fuer
Details.}
}
\value{
Regularisierter Tiqqle \code{x}.
}
\description{
Fuellt fuer jedes Signal evtl. vorhandende zeitliche Luecken auf. Dabei wird
fuer jeden im Dataframe auftretenden Zeitstempel ein entsprechender
Signalwert eingefuegt. Darueber hinaus koennen auch alle Luecken aufgefuellt
werden, die nur implizit vorliegen (d.h. Zeitstempel fuer die kein regulaerer
Signalwert existiert).
}
\seealso{
\code{\link{seq.POSIXt}}, \code{\link{fill}}

Other Zeitfunktionen: 
\code{\link{thicken}()}
}
\concept{Zeitfunktionen}
