% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{relative_time}
\alias{relative_time}
\title{Relative Zeitpunkte}
\usage{
relative_time(
  x,
  units = "secs",
  origin = lubridate::origin,
  as_datetime = TRUE
)
}
\arguments{
\item{x}{POSIXct-Vektor mit den Zeitpunkten.}

\item{units}{String mit der zu verwendenden Zeiteinheit (Default:
\emph{"secs"}, d.h. Sekunden).}

\item{as_datetime}{Logischer Wert, ob Ergebnis als Datetime (Ursprung:
1970-01-01) zurueckgegeben wird (Default: \emph{TRUE}).}

\item{POSIXct-Zeitstempel}{des Ursprungs (= Nullpunkt) (Default:
\{emph{lubridate::origin}).}
}
\value{
Vektor mit den relativen Zeitpunkten.
}
\description{
Bereinigt Zeitpunkte um die absolute Zeitinformation. Dabei erhaelt der erste
Zeitpunkt den Wert 0 und der zweite Zeitpunkt dann entsprechend den Offset
zum ersten Zeitpunkt usw.
}
\seealso{
\code{\link{origin}}
}
\keyword{internal}
