% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_constant}
\alias{remove_constant}
\title{Ohne konstante Signale}
\usage{
remove_constant(x, ...)
}
\arguments{
\item{x}{Tiqqle, der gefiltert wird.}

\item{...}{Weitere Argumente, die an Funktion \code{\link{is_constant}}
durchgereicht werden.}
}
\value{
Tiqqle \code{x} ohne konstante Signale.
}
\description{
Filtert konstante Signale aus einem Dataframe aus. Konstant heisst dabei,
dass im gesamten Zeitraum nur ein einziger Signalwert auftritt.
}
\seealso{
\code{\link{is_constant}}

Other Reduktionsfunktionen: 
\code{\link{condense}()},
\code{\link{normalize}()}
}
\concept{Reduktionsfunktionen}
