% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rename2}
\alias{rename2}
\title{Spaltenumbenennung}
\usage{
rename2(x, pattern_prefix = "", prefix = "", pattern_suffix = "", suffix = "")
}
\arguments{
\item{x}{Dataframe, der umbenannt wird.}

\item{pattern_prefix}{String mit regulaerem Ausdruck zur Eingrenzung der
Prefix-Position (Default: \emph{""}, d.h. es wird nichts eingegrenzt).}

\item{prefix}{String, der am Anfang der neuen Spaltennamen steht (Default:
\emph{""}, d.h. es wird nichts vorangestellt).}

\item{pattern_suffix}{String mit regulaerem Ausdruck zur Eingrenzung der
Suffix-Position (Default: \emph{""}, d.h. es wird nichts eingegrenzt).}

\item{suffix}{String, der am Ende der neuen Spaltennamen steht (Default:
\emph{""}, d.h. es wird nichts angehaengt).}
}
\value{
Dataframe mit umbenannten Spalten.
}
\description{
Benennt alle Spalten eines Dataframes um.
}
\keyword{internal}
