% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{seq_event}
\alias{seq_event}
\title{Alle Ereignisse}
\usage{
seq_event(
  int_start,
  int_end,
  event_length_in_sec,
  event_overlap_in_sec = 0,
  from_start_to_end = TRUE
)
}
\arguments{
\item{int_start}{POSIXct-Zeitstempel mit Startzeitpunkt des Sample-Zeitraums.}

\item{int_end}{POSIXct-Zeitstempel mit Endzeitpunkt des Sample-Zeitraums.}

\item{event_length_in_sec}{Numerischer Wert mit der Laenge der Ereignisse (in
Sekunden).}

\item{event_overlap_in_sec}{Numerischer Wert mit der Laenge des Intervalls
(in Sekunden), in dem sich zwei aufeinanderfolgende Ereignisse ueberlappen
duerfen (Default: 0, d.h. keine Ueberlappung erlaubt).}

\item{from_start_to_end}{Logischer Wert, ob Ereignisse von \code{int_start}
nach \code{int_end} erzeugt werden oder andersrum (Default: \emph{TRUE},
d.h. es wird bei \code{int_start} begonnen).}
}
\value{
Dataframe mit allen Ereignissen im Zeitraum.
}
\description{
Hilfsfunktion zum sequentiellen Erzeugen aller moeglichen Ereignisse in einem
gegebenen Zeitraum. Dabei kann neben dem Start- und Endzeitpunkt die
Ereignislaenge und die Ueberlappung sowie die Richtung der Sequenz veraendert
werden.
}
\keyword{internal}
