% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{thicken}
\alias{thicken}
\title{Aggregierte zeitliche Signalaufloesung}
\usage{
thicken(x, interval_sec = 1, fun = dplyr::first, ...)
}
\arguments{
\item{x}{Tiqqle, der zeitlich aggregiert wird.}

\item{interval_sec}{Numerischer Wert mit der Einheit (in Sekunden) der neuen
Zeitaufloesung (Default: 1, d.h. Signale werden in 1-Sekunden-Intervalle
aggregiert).}

\item{fun}{Funktion, die zur Aggregation der Signalwerte verwendet wird
(Default: \code{dplyr::first}, d.h. der erste im Intervall auftretende
Signalwert wird beibehalten).}

\item{...}{Weitere Argumente, die an Funktion \code{fun} durchgereicht
werden.}
}
\value{
Aggregierter Tiqqle \code{x}.
}
\description{
Fasst Signalwerte in einem hoeheren zeitlichen Intervall zusammen. Dabei
werden die urspruenglichen Zeitstempel zunaechst einer hoeheren
Zeitaufloesung zugeordnet. Anschliessend werden Signalwerte mit demselben
(neuen) Zeitstempel mittels einer Aggregationsfunktion zusammengefasst.
Standardmaessig wird hier der erste Signalwert beibehalten.
}
\details{
Prinzipiell laesst sich hierfuer auch eine andere Aggregationsfunktion
einsetzen (z.B. Mittelwertbildung). Diese vorher zu definierende Funktion
bekommt als erstes Argument den numerischen Vektor der Signalwerte uebergeben
und gibt einen entsprechend aggregierten numerischen Signalwert zurueck.
}
\seealso{
Other Zeitfunktionen: 
\code{\link{regularize}()}
}
\concept{Zeitfunktionen}
