% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{thicken_prepare}
\alias{thicken_prepare}
\title{Vorbereitung zeitlicher Signalaggregation}
\usage{
thicken_prepare(x, interval_sec = 1)
}
\arguments{
\item{x}{Tiqqle, der zeitlich aggregiert wird.}

\item{interval_sec}{Numerischer Wert mit der Einheit (in Sekunden) der neuen
Zeitaufloesung (Default: 1, d.h. Signale werden in 1-Sekunden-Intervalle
aggregiert).}
}
\value{
Zur Aggregation vorbereiteter Tiqqle \code{x}.
}
\description{
Hilfsfunktion zur Vorbereitung der zeitlichen Signalaggregation
(\code{\link{thicken}}).
}
\seealso{
\code{\link{thicken}}
}
\keyword{internal}
