% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiqqle.R
\name{tiqqle}
\alias{tiqqle}
\title{Tiqqle: ein Tibble}
\usage{
tiqqle(x = NULL, format = "long")
}
\arguments{
\item{x}{Dataframe, der konvertiert wird (Default: \emph{NULL}, d.h. es wird
ein leerer Tiqqle erzeugt).}

\item{format}{String, welches Format der erzeugte Tiqqle hat. Erlaubte Werte:
\emph{"long"} oder \emph{"wide"} (Default: \emph{"long"}, d.h. langes
Format).}
}
\value{
Tiqqle im gewuenschten Format.
}
\description{
Erzeugt einen Tiqqle-Dataframe im langen oder breiten Format. Falls kein
bereits existierender Dataframe uebergeben wird, wird ein leerer Tiqqle
erzeugt. Ansonsten wird er mit \code{as_tiqqle} konvertiert.
}
\seealso{
Other Tiqqle-Funktionen: 
\code{\link{as_long}()},
\code{\link{as_tiqqle}()},
\code{\link{as_wide}()}
}
\concept{Tiqqle-Funktionen}
