% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{value_has_changed}
\alias{value_has_changed}
\title{Geaenderte Werte?}
\usage{
value_has_changed(x)
}
\arguments{
\item{x}{Dataframe oder Vektor, der ueberprueft wird.}
}
\value{
Logischer Vektor, in welchen Zeilen von \code{x} Wertaenderungen
  auftreten.
}
\description{
Ueberprueft, in welchen Zeilen eines Dataframes oder Vektors sich die Werte
veraendert haben. Dabei wird spaltenweise die Differenz zweier
aufeinanderfolgender Werte betrachtet. Abschliessend werden die spaltenweisen
Ergebnisse zeilenweise zusammengefasst.
}
\keyword{internal}
