% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_align}
\alias{window_align}
\title{Fensterausrichtung}
\usage{
window_align(x, .progress = FALSE, ...)
}
\arguments{
\item{x}{Liste mit Fenster-Tiqqles, die zeitlich ausgerichtet werden.}

\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}

\item{...}{Weitere Argumente, die an \code{\link{relative_time}}
durchgereicht werden.}
}
\value{
Liste mit zeitlich ausgerichteten Fenster-Tiqqles.
}
\description{
Eine Menge von Fenster-Dataframes wird auf eine gemeinsame zeitliche Achse
gebracht. Dabei wird fuer jedes Fenster die absolute Zeitinformation in eine
relative ueberfuehrt.
}
\details{
Von allen Zeitstempeln wird der kleinste vorliegende Zeitstempel abgezogen
und somit ein neuer (numerischer) Ursprungszeitpunkt 0 geschaffen. Bei Bedarf
laesst sich dieser wiederum in einen POSIXct-Zeitstempel mit Ursprung
1970-01-01 umwandeln.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

\seealso{
\code{\link{relative_time}}

Other Fensterfunktionen: 
\code{\link{window_crop}()},
\code{\link{window_merge}()},
\code{\link{window_stretch}()},
\code{\link{window_truncate}()}
}
\concept{Fensterfunktionen}
