% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_crop}
\alias{window_crop}
\title{Fenster aus Signalen}
\usage{
window_crop(x, start, end, fun = NULL, .progress = TRUE, ...)
}
\arguments{
\item{x}{Tiqqle aus dem die Fenster extrahiert werden.}

\item{start}{POSIXct-Vektor mit Startzeitpunkten der Fenster.}

\item{end}{POSIXct-Vektor mit Endzeitpunkten der Fenster.}

\item{fun}{Funktion, die direkt nach der Extraktion zusaetzlich auf die
resultierende Fenster angewendet wird (Default: \emph{NULL}, d.h. es wird
keine weitere Funktion ausgefuehrt).}

\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}

\item{...}{Weitere Argumente, die ggf. an Funktion \code{fun} durchgereicht
werden.}
}
\value{
Liste mit extrahierten Fenster-Tiqqles.
}
\description{
Schneidet fuer eine Ereignis-Menge die zugehoerigen Signale aus einem
Signal-Dataframe. Dabei werden die Ereignisse durch ihre paarweisen Start-
und Endzeitpunkte definiert. Diese Funktion ist ein Wrapper um \code{crop}.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

\seealso{
crop

Other Fensterfunktionen: 
\code{\link{window_align}()},
\code{\link{window_merge}()},
\code{\link{window_stretch}()},
\code{\link{window_truncate}()}
}
\concept{Fensterfunktionen}
