% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_merge}
\alias{window_merge}
\title{Fensterverschmelzung}
\usage{
window_merge(x, fun = mean, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Liste mit Fenster-Tiqqles, die verschmolzen werden.}

\item{fun}{Funktion, die zum Verschmelzen der Fenster verwendet wird
(Default: \code{mean}, d.h. Mittelwertbildung).}

\item{na.rm}{Logischer Wert, ob vorhandene \emph{NA}-Werte ignoriert werden
(Default: \emph{TRUE}, d.h. sie werden nicht beruecksichtigt).}

\item{...}{Weitere Argumente, die an Funktion \code{fun} durchgereicht
werden.}
}
\value{
Dataframe mit verschmolzenen Fenster-Tiqqles.
}
\description{
Eine Menge von Fenster-Dataframes wird zu einem einzigen Fenster-Dataframe
zusammengefasst. Dabei werden die Signalwerte der Einzelfenster punktweise
(d.h. jeweils identische Signale und Zeitpunkte) mittels einer
Verschmelzungsfunktion (z.B. Mittelwertbildung) aggregiert.
}
\details{
Prinzipiell laesst sich hierfuer auch eine andere Verschmelzungsfunktion
einsetzen. Diese vorher zu definierende Funktion bekommt als erstes Argument
den numerischen Vektor der punktweisen Signalwerte uebergeben und gibt einen
einzelnen numerischen Rueckgabewert zurueck.
}
\seealso{
Other Fensterfunktionen: 
\code{\link{window_align}()},
\code{\link{window_crop}()},
\code{\link{window_stretch}()},
\code{\link{window_truncate}()}
}
\concept{Fensterfunktionen}
