% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_merge_aggregate}
\alias{window_merge_aggregate}
\alias{window_merge_aggregate.tiqqle_long}
\alias{window_merge_aggregate.tiqqle_wide}
\title{Fensterzusammenfassung}
\usage{
window_merge_aggregate(x, fun, na.rm, ...)

\method{window_merge_aggregate}{tiqqle_long}(x, fun, na.rm, ...)

\method{window_merge_aggregate}{tiqqle_wide}(x, fun, na.rm, ...)
}
\arguments{
\item{x}{Fenster-Tiqqle, der verschmolzen wird.}

\item{fun}{Funktion, die zum Verschmelzen der Fenster verwendet wird.}

\item{na.rm}{Logischer Wert, ob vorhandene \emph{NA}-Werte ignoriert werden.}

\item{...}{Weitere Argumente, die an Funktion \code{fun} durchgereicht
werden.}
}
\value{
Dataframe mit zusammengefassten Fenstern.
}
\description{
Hilfsfunktion zum Verschmelzen von Fenstern (\code{\link{window_merge}}). Ein
Fenster-Dataframe wird zusammengefasst und evtl. fehlende Zeitpunkte mit
entsprechenden Signalwerten aufgefuellt.
}
\section{Methods (by class)}{
\itemize{
\item \code{tiqqle_long}: Fasst Fenster im langen Format beim
Verschmelzen zusammen

\item \code{tiqqle_wide}: Fasst Fenster im breiten Format beim
Verschmelzen zusammen
}}

\seealso{
\code{\link{window_merge}}
}
\keyword{internal}
