% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_truncate}
\alias{window_truncate}
\title{Fensterbeschneidung}
\usage{
window_truncate(x, .progress = FALSE)
}
\arguments{
\item{x}{Liste mit Fenster-Tiqqles, die zeitlich beschnitten werden.}

\item{.progress}{Logischer Wert, ob Fortschrittsbalken angezeigt wird.}
}
\value{
Liste mit zeitlich beschnittenen Fenster-Tiqqles.
}
\description{
Eine Menge von Fenster-Dataframes wird auf eine einheitliche Laenge gebracht.
Dabei wird zunaechst das kuerzeste Fenster bestimmt. Dessen Laenge wird dann
auf die restlichen Fenster angewendet und evtl. vorhandene Signalaenderungen
nach diesem Zeitpunkt verworfen. Bei diesem Ansatz gehen also
Signalinformationen verloren.
}
\section{Future}{
 Da dieser Verarbeitungsschritt u.U. etwas laenger dauern
  kann, kann hierfuer mittels \code{future::plan} eine parallele Verarbeitung
  zugeschaltet werden. Durch das Argument \code{.progress} kann ggf. ein
  Forschrittsbalken angezeigt werden.
}

\seealso{
Other Fensterfunktionen: 
\code{\link{window_align}()},
\code{\link{window_crop}()},
\code{\link{window_merge}()},
\code{\link{window_stretch}()}
}
\concept{Fensterfunktionen}
