% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_tiqqle}
\alias{is_tiqqle}
\title{Tiqqle?}
\usage{
is_tiqqle(x)
}
\arguments{
\item{x}{Dataframe, dessen Format ueberprueft wird.}
}
\value{
Logischer Wert, ob Dataframe \code{x} ein Tiqqle ist.
}
\description{
Ueberprueft, ob ein Dataframe ein langer oder breiter Tiqqle ist.
}
\seealso{
Other Format-Checks: 
\code{\link{is_long}()},
\code{\link{is_valid}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
