% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_valid}
\alias{is_valid}
\title{Valider Tiqqle?}
\usage{
is_valid(x)
}
\arguments{
\item{x}{Dataframe (\code{\link{tibble}}), dessen Format ueberprueft wird.}
}
\value{
Logischer Wert, ob Dataframe \code{x} ein valider Tiqqle ist.
}
\description{
Ueberprueft, ob ein Dataframe ein langer oder breiter Tiqqle ist.
}
\seealso{
Other Format-Checks: 
\code{\link{is_long}()},
\code{\link{is_tiqqle}()},
\code{\link{is_wide}()}
}
\concept{Format-Checks}
